pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.13.2",
    "numpy>=2.2.6",
    "ollama>=0.6.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]

[project.scripts]
rag-system = "rag_system.main:main"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]